broke unit_tests into multiple files to match core files
