+++ test/player_test.py
@@ -1,0 +1,6 @@
+"""
+"""
+from base_test import CGTestCase
+
+class PlayerTests(CGTestCase):
+	pass
